Изменяю работу главного фрагмента, добавляю проверку на бд, синхронизацию с ней и отображение данных

+++ domain/src/main/java/com/example/domain/models/PokemonsCardsModel.kt
@@ -1,5 +1,7 @@
 package com.example.domain.models
 
+import android.annotation.SuppressLint
+import android.util.Log
 import com.example.mypokemons.data.database.room.PokemonEntity
 import com.example.mypokemons.data.database.room.models.Images
 import com.example.mypokemons.data.di.AppComponent
@@ -7,6 +9,7 @@ import com.example.mypokemons.data.storage.PokemonInfo
 import com.example.mypokemons.data.storage.PokemonModel
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
+import java.lang.Exception
 
 
 class PokemonsCardsModel(val appComponent: AppComponent) {
@@ -18,45 +21,36 @@ class PokemonsCardsModel(val appComponent: AppComponent) {
 
     fun getPreviewRoomCards(): Single<List<PokemonEntity>> = pokemonsDao.getAll()
 
-    fun refreshTable(insertData: List<PokemonInfo>) {
-        pokemonsDao.getSize()
-            .subscribeOn(Schedulers.io())
-            .subscribe { count ->
-                //if (count > 0)
-                    pokemonsDao.deleteAll()
+    fun refreshTable(insertData: List<PokemonInfo>) =
+        pokemonsDao.getAll()
+            .map { dbPokemons ->
+                if(insertData.isEmpty())
+                    return@map dbPokemons
 
-                val dbList = insertData.map {
-                    PokemonEntity(
-                        0,
-                        it.name,
-                        "",
-                        "it.types.first()",
-                        "it.subtypes.first()",
-                        it.hp,
-                        Images(it.images.small, it.images.large),
-                        "it.attacks.first().name"
-                    )
+                pokemonsDao.deleteAll()
+
+                val dbList = insertData.mapIndexed { index,pokemon ->
+                    var isFavorite = false
+                    if(dbPokemons.isNotEmpty() && dbPokemons.size >= index)
+                        isFavorite = dbPokemons[index].isFavorite
+
+                    var attack = "Nope"
+                    if(pokemon.attacks != null)
+                        attack = pokemon.attacks[0].name
+
+                        PokemonEntity(
+                            pokemon.name,
+                            pokemon.rarity ?: "",
+                            pokemon.types[0] ?: "",
+                            pokemon.subtypes[0] ?: "",
+                            pokemon.hp,
+                            Images(pokemon.images.small, pokemon.images.large),
+                            attack?: "",
+                            isFavorite
+                        )
                 }
                 pokemonsDao.insertList(dbList)
-            }
-    }
-    /*{
-        var pokemonsList: List<BasePokemonModel> = listOf()
-        apiService.getAllCards()
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribeOn(Schedulers.io())
-           .subscribe {
-               if(it.isSuccessful)
-               pokemonsList = BasePokemonModel.cast(it.body().pokemons){ pokemonInfo ->
-                   BasePokemonModel(pokemonInfo.name, pokemonInfo.images.small)
-               }
-               else{
-                   pokemonsList = BasePokemonModel.cast(pokemonsDao.getAll()){pokemonInfo ->
-                       BasePokemonModel(pokemonInfo.name, pokemonInfo.images.small)
-                   }
-               }
-           }
-        return Observable.just(pokemonsList)
-    }*/
 
+                return@map dbList
+            }
 }
